--- src/main/java/com/planup/planup/domain/goal/entity/TimerVerification.java
@@ -1,38 +1,0 @@
-package com.planup.planup.domain.goal.entity;
-
-import com.planup.planup.domain.global.entity.BaseTimeEntity;
-import com.planup.planup.domain.goal.entity.Enum.GoalPeriod;
-import com.planup.planup.domain.goal.entity.mapping.UserGoal;
-import jakarta.persistence.*;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
-
-import java.sql.Time;
-
-@Entity
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@SuperBuilder
-public class TimerVerification extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private Time goalTime;
-
-    private int frequency;
-
-    @Enumerated(EnumType.STRING)
-    private GoalPeriod period;
-
-    private Time spentTime;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "community_id")
-    private UserGoal community;
-}
