--- src/main/java/com/planup/planup/domain/user/verification/dto/TimerVerificationRequestDto.java
@@ -1,4 +1,0 @@
-package com.planup.planup.domain.user.verification.dto;
-
-public class TimerVerificationRequestDto {
-}
