--- src/main/java/com/planup/planup/domain/user/verification/entity/PhotoVerification.java
@@ -1,27 +1,0 @@
-package com.planup.planup.domain.user.verification.entity;
-
-import com.planup.planup.domain.global.entity.BaseTimeEntity;
-import com.planup.planup.domain.goal.entity.mapping.UserGoal;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
-
-@Entity
-@Getter
-@SuperBuilder
-@NoArgsConstructor
-@AllArgsConstructor
-public class PhotoVerification extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String photoImg;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "usergoal_id")
-    private UserGoal userGoal;
-}
