--- src/main/java/com/planup/planup/domain/user/verification/service/VerificationService.java
@@ -1,4 +1,0 @@
-package com.planup.planup.domain.user.verification.service;
-
-public interface VerificationService {
-}
