+++ src/main/java/com/planup/planup/domain/verification/repository/PhotoVerificationRepository.java
@@ -1,6 +1,6 @@
-package com.planup.planup.domain.verification.repository;
+package com.planup.planup.domain.user.verification.repository;
 
-import com.planup.planup.domain.verification.entity.PhotoVerification;
+import com.planup.planup.domain.user.verification.entity.PhotoVerification;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
