+++ src/main/java/com/planup/planup/domain/verification/service/VerificationService.java
@@ -1,4 +1,4 @@
-package com.planup.planup.domain.verification.service;
+package com.planup.planup.domain.user.verification.service;
 
 public interface VerificationService {
 }
